moved schemas into schema dir
